Low power test pattern generator for BIST architecture
Code of the low power test pattern generator for BIST architecture using
the finite state model as a part of the implementation of the research
paper titled "Low power test pattern generator for BIST architecture",
N. Ahmed, M.H. Tehranipour, M. Nourani. It was successfully implemented
with power results tested on Xilinx platform. The code was developed in
Questa Sim.